feat: lista de origen en vista @all, ciclo Pomodoro completo y robustez
Funcionalidades:
- Columna "Lista" en la vista @all (tabla, panel de info y modo enfoque)
  con helpers App::is_all_view() / App::list_title_for().
- Descanso largo automático cada 4 pomodoros + tecla 'm' para cambiar de
  modo manualmente (App::cycle_timer_mode).

Correcciones:
- Evita pánico (división por cero / index OOB) al navegar listas cuando
  task_lists está vacío tras un fallo de red.
- record_task_done() ahora cuenta al completar con o sin timer activo.
- Sin captura de mouse: se preserva la selección/copiado del terminal.
- Elimina eprintln! que corrompía la TUI en modo raw.

Calidad:
- build.rs ya no hace panic! sin client_secret.json: embebe credenciales
  (o placeholder) vía OUT_DIR, así CI y clones limpios compilan.
- Cargo.lock versionado (sale del .gitignore).
- Tests unitarios de parse_due_date y organize_tasks_hierarchical.
- 0 warnings de clippy.

Docs:
- AGENTS.md y CLAUDE.md para agentes de IA; README actualizado.

+++ src/ui/calendar.rs
@@ -114,7 +114,7 @@ pub fn render_calendar(app: &App, frame: &mut Frame, area: Rect) {
                 let mut cells = vec![Cell::from(format!("{:02}:00", h)).style(Style::default().fg(Palette::overlay0(app)))];
                 for i in 0..7 {
                     let date = start_of_week + chrono::Duration::days(i as i64);
-                    let count = hourly_done.get(&(date, h as u32)).unwrap_or(&0);
+                    let count = hourly_done.get(&(date, h)).unwrap_or(&0);
                     let symbol = if *count > 0 { "█" } else { "·" };
                     let style = if *count > 0 { Style::default().fg(Palette::green(app)) } else { Style::default().fg(Palette::surface0(app)) };
                     cells.push(Cell::from(Line::from(symbol).alignment(Alignment::Center)).style(style));
@@ -132,7 +132,7 @@ pub fn render_calendar(app: &App, frame: &mut Frame, area: Rect) {
             let start_h = current_hour.saturating_sub(5).min(14);
             
             for h in start_h..start_h+10 {
-                let count = hourly_done.get(&(date, h as u32)).unwrap_or(&0);
+                let count = hourly_done.get(&(date, h)).unwrap_or(&0);
                 let mut spans = vec![
                     Span::styled(format!("{:02}:00  ", h), Style::default().fg(Palette::subtext0(app))),
                 ];
